Add tests for compression, and a CLI tool for manual testing.
I changed the names of the tests, so that compress round-trip runs
after the uncompress tests.

Also added a cli tool that tries to follow the official /bro/ cli tool,
but it's use is mostly for validation and testing, and is still missing
support for many of the flags (eg. --repeat).

For this, I moved the File::Slurp module to an actual dependency, not
a tests-only dep.